Some clean-up of fixes

+++ scipy/special/special_wrappers.cpp
@@ -275,5 +275,5 @@ npy_cdouble special_sph_harm(long m, long n, double theta, double phi) {
 }
 
 npy_cdouble special_sph_harm_unsafe(double m, double n, double theta, double phi) {
-    return to_ccomplex(special::sph_harm(m, n, theta, phi));
+    return to_ccomplex(special::sph_harm(static_cast<long>(m), static_cast<long>(n), theta, phi));
 }
